Migrate example to mono repo, rename i18n folder name
