+++ src/app/pages/i18n/services/bidirectional.service.ts
@@ -5,7 +5,7 @@ import { Injectable } from '@angular/core';
 })
 export class BidirectionalService {
     changeDirectionality(lang: string): void {
-        const body = document.querySelector('body');
+        const body = document.querySelector('.i18n-container');
         body.setAttribute('dir', this.isRTL(lang) ? 'rtl' : 'ltr');
     }
