+++ src/app/pages/i18n/translations/arabic.ts
