+++ src/app/pages/i18n/translations/chinese.ts
