+++ src/app/pages/i18n/translations/english.ts
