+++ src/app/pages/i18n/translations/french.ts
