+++ src/app/pages/i18n/translations/sample-translation.ts
