+++ src/app/pages/i18n/translations/spanish.ts
